feat(admin): extract AdminHubSurface monolith into section components + tab navigation
Break the 1500-line AdminHubSurface.jsx into five focused section files
(Overview, Accounts, Debugging, Content, Marketing) behind a horizontal
tab bar, making each admin domain independently testable and navigable.

Test fixes:
- SSR render helper now supports initialSection and all-sections mode
- Error-copy oracle allowlist entries moved to AdminDebuggingSection.jsx
- CSP inline-style budget and classification updated for 6 new files
- All 13 previously-failing tests pass against the new section architecture
